+++ packages/editor-api/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @docx-editor.dev/editor-api
 
+## 2.2.1
+
+### Patch Changes
+
+- Updated dependencies [35f6d04]
+  - @docx-editor.dev/core@2.2.1
+
+## 2.2.0
+
+### Patch Changes
+
+- Updated dependencies [3096225]
+- Updated dependencies [9c25492]
+- Updated dependencies [04c2379]
+- Updated dependencies [f0e4ab9]
+  - @docx-editor.dev/core@2.2.0
+
 ## 2.1.3
 
 ### Patch Changes