+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @docx-editor.dev/nuxt
 
+## 2.2.1
+
+### Patch Changes
+
+- @docx-editor.dev/vue@2.2.1
+
+## 2.2.0
+
+### Patch Changes
+
+- @docx-editor.dev/vue@2.2.0
+
 ## 2.1.3
 
 ### Patch Changes